added topological distance variable

+++ wikibrain-spatial/src/main/java/org/wikibrain/spatial/maxima/SpatialConceptPair.java
@@ -54,7 +54,7 @@ public class SpatialConceptPair {
         return kmDistance;
     }
 
-    public double getTopDistance() {
+    public int getTopDistance() {
         return topDistance;
     }
